alcf model: Consider gaps in track
Add an option track_gap. This specifies the minimum time interval between two
track points which is considered a gap. Previously, location was always
interpolated between adjacent points. With this new option and its default
value of 6 hours, gaps are identified in the track for which no output is
produced. If time_bnds is present in the track, this option is ignored and no
gaps are detected other than already specified by time_bnds.